WIP DBA Refactoring
