+++ src/Database/Connection/ConnectionException.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Friendica\Database\Connection;
+
+use Exception;
+
+/**
+ * A DB connection exception
+ */
+class ConnectionException extends Exception
+{
+
+}
